Added DPS3003 default parameters and fixed overflow in pwrctl_calc_ilimit_adc. (#226)
The calibration of DPS3003 often resulted in malfunctioning of
CC and CL function due to a bug in pwrctl.c:pwrctl_calc_ilimit_adc:
The float value in pwrctl_calc_ilimit_adc might result in overflow
during the casting to the return value of type uint16_t.
The return type is changed to uint32_t.

+++ opendps/pwrctl.c
@@ -296,7 +296,7 @@ uint32_t pwrctl_calc_iout(uint16_t raw)
   * @param i_limit_ma selected I_limit
   * @retval expected raw ADC value
   */
-uint16_t pwrctl_calc_ilimit_adc(uint16_t i_limit_ma)
+uint32_t pwrctl_calc_ilimit_adc(uint16_t i_limit_ma)
 {
     float value = (i_limit_ma - a_adc_c_coef) / a_adc_k_coef + 1;
     if (value <= 0)
@@ -310,7 +310,7 @@ uint16_t pwrctl_calc_ilimit_adc(uint16_t i_limit_ma)
   * @param v_limit_mv selected V_limit
   * @retval expected raw ADC value
   */
-uint16_t pwrctl_calc_vlimit_adc(uint16_t v_limit_mv)
+uint32_t pwrctl_calc_vlimit_adc(uint16_t v_limit_mv)
 {
     float value = (v_limit_mv - v_adc_c_coef) / v_adc_k_coef + 1;
     if (value <= 0)
